[ADDED] Validate input via useEffect
The validation of the input entered by the user takes place through a
useEffect that inhibits the insertion of negative values and limits the
maximum insearable number to the length of the list. This is a personal
choice, different from the tutorial solution which suggests a validation
on the value obtained from the input and not on the input itself.