[UPDATED] in context.js useCallback in fetchDrinks
The logic of the fetchDrinks function is contained in useCallback, which
takes the hooked variable searchTerm as its second parameter. At the
same time, the reference to the function itself and the hooked variable
searchTerm are passed as the second argument to the useEffect that
invokes the fetchDrinks function. This way multiple fetch requests on
searchTerm value change will work correctly, without loops, and React
does not report the lack of dependencies for useEffect.